+++ src/architectures/neural_network_integrator.jl
@@ -73,7 +73,7 @@ using GeometricMachineLearning
 model = ResNet(3, 0, identity)
 weight = [1 0 0; 0 2 0; 0 0 1]
 bias = [0, 0, 1]
-ps = NeuralNetworkParameters((L1 = (weight = weight, bias = bias), ))
+ps = NetworkParameters((L1 = (weight = weight, bias = bias), ))
 nn = NeuralNetwork(model, Chain(model), ps, CPU())
 
 ics = [1, 1, 1]